correccion errores compilacion
