+++ src/containers/CheckoutPage/CheckoutPage.js
@@ -35,7 +35,7 @@ import {
 import CustomTopbar from './CustomTopbar';
 import CheckoutPageWithPayment, {
   loadInitialData,
-} from './CheckoutPageWithPayment';
+} from './CheckoutPageWithoutPayment';
 import CheckoutPageWithInquiryProcess from './CheckoutPageWithInquiryProcess';
 
 const STORAGE_KEY = 'CheckoutPage';
